#193: per-OS FilePicker impls + factory, single OS seam, platform.* packages, DI cleanup
