+++ composeApp/src/commonTest/kotlin/com/tubetoast/tether/transfer/PendingFilesRepositoryTest.kt
@@ -8,15 +8,13 @@ import kotlin.test.assertNull
 import kotlin.test.assertTrue
 
 class PendingFilesRepositoryTest {
-    private val firstSummary = PendingFilesSummary(1, 100L)
-    private val secondSummary = PendingFilesSummary(1, 200L)
     private val first = listOf(FakeFileSource("first.txt", 100L))
     private val second = listOf(FakeFileSource("second.txt", 200L))
 
     @Test
     fun `clear unconditionally drops pending`() {
         val repo = PendingFilesRepository()
-        repo.setPending(firstSummary, first)
+        repo.setPending(first)
         repo.clear()
         assertNull(repo.pending.value)
     }
@@ -24,7 +22,7 @@ class PendingFilesRepositoryTest {
     @Test
     fun `clearIfMatches returns true and clears when token matches current snapshot`() {
         val repo = PendingFilesRepository()
-        repo.setPending(firstSummary, first)
+        repo.setPending(first)
         val token = repo.pending.value!!
         assertTrue(repo.clearIfMatches(token))
         assertNull(repo.pending.value)
@@ -33,10 +31,10 @@ class PendingFilesRepositoryTest {
     @Test
     fun `clearIfMatches returns false and preserves state when a fresh setPending raced ahead`() {
         val repo = PendingFilesRepository()
-        repo.setPending(firstSummary, first)
+        repo.setPending(first)
         val token = repo.pending.value!!
         // Simulate the race: the caller captured `token` but a new share landed before clearIfMatches.
-        repo.setPending(secondSummary, second)
+        repo.setPending(second)
         assertFalse(repo.clearIfMatches(token))
         assertEquals(second, repo.pending.value?.sources)
         assertNotNull(repo.pending.value?.summary)
@@ -45,10 +43,10 @@ class PendingFilesRepositoryTest {
     @Test
     fun `setPending atomically pairs summary and sources`() {
         val repo = PendingFilesRepository()
-        repo.setPending(firstSummary, first)
+        repo.setPending(first)
         val snapshot = repo.pending.value
         assertNotNull(snapshot)
-        assertEquals(firstSummary, snapshot.summary)
+        assertEquals(PendingFilesSummary.from(first), snapshot.summary)
         assertEquals(first, snapshot.sources)
     }
 }
