+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/platform/win/GuidFixed.kt
@@ -1,6 +1,6 @@
 // Adapted from FileKit (MIT, github.com/vinceglb/FileKit).
 
-package com.tubetoast.tether.transfer.win
+package com.tubetoast.tether.platform.win
 
 import com.sun.jna.platform.win32.Guid
