+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/platform/win/IFileOpenDialog.kt
@@ -6,7 +6,7 @@
     "ConstPropertyName",
 )
 
-package com.tubetoast.tether.transfer.win
+package com.tubetoast.tether.platform.win
 
 import com.sun.jna.platform.win32.WinNT
 import com.sun.jna.ptr.PointerByReference
