+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/platform/win/IShellItemArray.kt
@@ -6,7 +6,7 @@
     "ConstPropertyName",
 )
 
-package com.tubetoast.tether.transfer.win
+package com.tubetoast.tether.platform.win
 
 import com.sun.jna.Pointer
 import com.sun.jna.platform.win32.COM.IUnknown
@@ -14,8 +14,8 @@ import com.sun.jna.platform.win32.Guid
 import com.sun.jna.platform.win32.WinNT
 import com.sun.jna.ptr.IntByReference
 import com.sun.jna.ptr.PointerByReference
-import com.tubetoast.tether.transfer.win.ShTypes.GETPROPERTYSTOREFLAGS
-import com.tubetoast.tether.transfer.win.ShTypes.PROPERTYKEY
+import com.tubetoast.tether.platform.win.ShTypes.GETPROPERTYSTOREFLAGS
+import com.tubetoast.tether.platform.win.ShTypes.PROPERTYKEY
 
 internal interface IShellItemArray : IUnknown {
     fun BindToHandler(
