+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/platform/win/ShTypes.kt
@@ -6,7 +6,7 @@
     "ConstPropertyName",
 )
 
-package com.tubetoast.tether.transfer.win
+package com.tubetoast.tether.platform.win
 
 import com.sun.jna.Structure
 import com.sun.jna.Structure.FieldOrder
