+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/platform/win/ShellItem.kt
@@ -6,7 +6,7 @@
     "ConstPropertyName",
 )
 
-package com.tubetoast.tether.transfer.win
+package com.tubetoast.tether.platform.win
 
 import com.sun.jna.Pointer
 import com.sun.jna.platform.win32.COM.Unknown
@@ -14,12 +14,10 @@ import com.sun.jna.platform.win32.Guid
 import com.sun.jna.platform.win32.WinNT
 import com.sun.jna.ptr.IntByReference
 import com.sun.jna.ptr.PointerByReference
-import com.tubetoast.tether.transfer.win.ShTypes.GETPROPERTYSTOREFLAGS
-import com.tubetoast.tether.transfer.win.ShTypes.PROPERTYKEY
 
-internal class ShellItemArray :
+internal class ShellItem :
     Unknown,
-    IShellItemArray {
+    IShellItem {
     constructor()
 
     constructor(pvInstance: Pointer?) : super(pvInstance)
@@ -29,52 +27,44 @@ internal class ShellItemArray :
         pbc: Pointer?,
         bhid: Guid.GUID.ByReference?,
         riid: Guid.REFIID?,
-        ppvOut: PointerByReference?,
+        ppv: PointerByReference?,
     ): WinNT.HRESULT = _invokeNativeObject(
         3,
-        arrayOf(this.pointer, pbc, bhid, riid, ppvOut),
+        arrayOf(this.pointer, pbc, bhid, riid, ppv),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
-    override fun GetPropertyStore(
-        flags: GETPROPERTYSTOREFLAGS?,
-        riid: Guid.REFIID?,
-        ppv: PointerByReference?,
-    ): WinNT.HRESULT = _invokeNativeObject(
+    override fun GetParent(ppsi: PointerByReference?): WinNT.HRESULT = _invokeNativeObject(
         4,
-        arrayOf(this.pointer, flags, riid, ppv),
+        arrayOf(this.pointer, ppsi),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
-    override fun GetPropertyDescriptionList(
-        keyType: PROPERTYKEY?,
-        riid: Guid.REFIID?,
-        ppv: PointerByReference?,
+    override fun GetDisplayName(
+        sigdnName: Long,
+        ppszName: PointerByReference?,
     ): WinNT.HRESULT = _invokeNativeObject(
         5,
-        arrayOf(this.pointer, keyType, riid, ppv),
+        arrayOf(this.pointer, sigdnName, ppszName),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
     override fun GetAttributes(
-        AttribFlags: Int,
         sfgaoMask: Int,
         psfgaoAttribs: IntByReference?,
     ): WinNT.HRESULT = _invokeNativeObject(
         6,
-        arrayOf(this.pointer, AttribFlags, sfgaoMask, psfgaoAttribs),
+        arrayOf(this.pointer, sfgaoMask, psfgaoAttribs),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
-    override fun GetCount(pdwNumItems: IntByReference?): WinNT.HRESULT = _invokeNativeObject(
+    override fun Compare(
+        psi: Pointer?,
+        hint: Int,
+        piOrder: IntByReference?,
+    ): WinNT.HRESULT = _invokeNativeObject(
         7,
-        arrayOf(this.pointer, pdwNumItems),
-        WinNT.HRESULT::class.java,
-    ) as WinNT.HRESULT
-
-    override fun GetItemAt(dwIndex: Int, ppsi: PointerByReference?): WinNT.HRESULT = _invokeNativeObject(
-        8,
-        arrayOf(this.pointer, dwIndex, ppsi),
+        arrayOf(this.pointer, psi, hint, piOrder),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 }
