+++ composeApp/src/desktopMain/kotlin/com/tubetoast/tether/platform/win/ShellItemArray.kt
@@ -6,7 +6,7 @@
     "ConstPropertyName",
 )
 
-package com.tubetoast.tether.transfer.win
+package com.tubetoast.tether.platform.win
 
 import com.sun.jna.Pointer
 import com.sun.jna.platform.win32.COM.Unknown
@@ -14,10 +14,12 @@ import com.sun.jna.platform.win32.Guid
 import com.sun.jna.platform.win32.WinNT
 import com.sun.jna.ptr.IntByReference
 import com.sun.jna.ptr.PointerByReference
+import com.tubetoast.tether.platform.win.ShTypes.GETPROPERTYSTOREFLAGS
+import com.tubetoast.tether.platform.win.ShTypes.PROPERTYKEY
 
-internal class ShellItem :
+internal class ShellItemArray :
     Unknown,
-    IShellItem {
+    IShellItemArray {
     constructor()
 
     constructor(pvInstance: Pointer?) : super(pvInstance)
@@ -27,44 +29,52 @@ internal class ShellItem :
         pbc: Pointer?,
         bhid: Guid.GUID.ByReference?,
         riid: Guid.REFIID?,
-        ppv: PointerByReference?,
+        ppvOut: PointerByReference?,
     ): WinNT.HRESULT = _invokeNativeObject(
         3,
-        arrayOf(this.pointer, pbc, bhid, riid, ppv),
+        arrayOf(this.pointer, pbc, bhid, riid, ppvOut),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
-    override fun GetParent(ppsi: PointerByReference?): WinNT.HRESULT = _invokeNativeObject(
+    override fun GetPropertyStore(
+        flags: GETPROPERTYSTOREFLAGS?,
+        riid: Guid.REFIID?,
+        ppv: PointerByReference?,
+    ): WinNT.HRESULT = _invokeNativeObject(
         4,
-        arrayOf(this.pointer, ppsi),
+        arrayOf(this.pointer, flags, riid, ppv),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
-    override fun GetDisplayName(
-        sigdnName: Long,
-        ppszName: PointerByReference?,
+    override fun GetPropertyDescriptionList(
+        keyType: PROPERTYKEY?,
+        riid: Guid.REFIID?,
+        ppv: PointerByReference?,
     ): WinNT.HRESULT = _invokeNativeObject(
         5,
-        arrayOf(this.pointer, sigdnName, ppszName),
+        arrayOf(this.pointer, keyType, riid, ppv),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
     override fun GetAttributes(
+        AttribFlags: Int,
         sfgaoMask: Int,
         psfgaoAttribs: IntByReference?,
     ): WinNT.HRESULT = _invokeNativeObject(
         6,
-        arrayOf(this.pointer, sfgaoMask, psfgaoAttribs),
+        arrayOf(this.pointer, AttribFlags, sfgaoMask, psfgaoAttribs),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 
-    override fun Compare(
-        psi: Pointer?,
-        hint: Int,
-        piOrder: IntByReference?,
-    ): WinNT.HRESULT = _invokeNativeObject(
+    override fun GetCount(pdwNumItems: IntByReference?): WinNT.HRESULT = _invokeNativeObject(
         7,
-        arrayOf(this.pointer, psi, hint, piOrder),
+        arrayOf(this.pointer, pdwNumItems),
+        WinNT.HRESULT::class.java,
+    ) as WinNT.HRESULT
+
+    override fun GetItemAt(dwIndex: Int, ppsi: PointerByReference?): WinNT.HRESULT = _invokeNativeObject(
+        8,
+        arrayOf(this.pointer, dwIndex, ppsi),
         WinNT.HRESULT::class.java,
     ) as WinNT.HRESULT
 }
